Update readme.md; Reorganize the code structure
